// src/types/marketplace-types.d.ts
export type TabItemConfig = {
  name: string;
  enabled: boolean;
};